CHROMIUM: iwl7000: mvm: skip short statistics window when updating EMLSR
The statistics are not synchronized with the time that we enter EMLSR.
This means that we can receive the statistic notification just after
having cleared the counters, causing us to immediately exit EMLSR again.

Fix this by checking that most of the time for the window has passed. If
that is not the case, ignore this window and wait for the next
notification.

Change-Id: Ic2af92737ccfc873f3b6c228704238ebb9f983ca
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: a4bcfb6266ede007a7d6ccd8e8aa084788fae790