CHROMIUM: iwl7000: mld: add support for beacon filtering
Just send the command whenever we get the CQM configuration.
The firmware will not start filtering the beacons until a few seconds
after association, so no risk to miss the first beacons.

Don't disable beacon filtering upon interface addition.
This does nothing in the firmware if this is the first interface.
If this is the second, we'll enable beacon filter after its association
(but since this will  likely be a p2p client, we won't send any command
to the firmware since beacon filtering is not supported for p2p client).

Keep the disable beacon filtering function, it'll be used for debugfs
purposes.

Change-Id: I656a89375c687cea2af673d391d8796464896a88
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 66f0e31e6a3fb5385d912be79b88ad098025a123