UPSTREAM: arm64: dts: mediatek: mt8173-elm-hana: Add vdd-supply to second source trackpad
The Hana device has a second source option trackpad, but it is missing
its regulator supply. It only works because the regulator is marked as
always-on.

Add the regulator supply, but leave out the post-power-on delay. Instead,
document the post-power-on delay along with the reason for not adding
it in a comment.

Fixes: 689b937bedde ("arm64: dts: mediatek: add mt8173 elm and hana board")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20241018082001.1296963-1-wenst@chromium.org
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit f766fae08f6a2eaeb45d8d2c053724c91526835c)

BUG=b:313564065
TEST=cros build-packages -b kukui chromeos-kernel-6_6

Change-Id: Ib4b97d3a42c37d0e84d1ee4f594c09a55920374b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6076474
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Chen-Yu Tsai <wenst@chromium.org>