Revert "FROMLIST: dt-bindings: i2c: add attribute default-timing-adjust"
This reverts commit af209b301e754ed5cb77d0ba8b43507e6e91a1a9.

It appears that this patch was just carried forward for no good
reason. Archaeology shows:

The attribute was called "default-timing-adjust" in v4 of the series
[1] but got changed to "use-default-timing" in v5 [2]. The actual code
to use this seems to have lasted till v7 [3] and v8 [4]. It looks like
earlier patches in v8 landed but then it looks like Wolfram asked why
we couldn't just handle things more automatically. The conclusion [5]
was that it could. Then it looks like the attribute was dropped. The
final patch was then posted at v9 without needing the devicetree
attribute.

Grepping through the kernel sources neither "default-timing-adjust"
nor "use-default-timing" is used so it's pretty clear we can drop this
patch.

BUG=b:371946324
TEST=None

[1] https://lore.kernel.org/lkml/1626517079-9057-6-git-send-email-kewei.xu@mediatek.com/
[2] https://lore.kernel.org/lkml/1629723937-10839-5-git-send-email-kewei.xu@mediatek.com/
[3] https://lore.kernel.org/lkml/20210917101416.20760-7-kewei.xu@mediatek.com/
[4] https://lore.kernel.org/lkml/20211010070516.26763-5-kewei.xu@mediatek.com/
[5] https://lore.kernel.org/lkml/Yb21HBSm4VVjy1cr@kunai/

Change-Id: I2da69d5a5e31b183188794646c892ba714a68d28
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5941691
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>