Revert "CHROMIUM: arm64: mt8186: dts: add proprietary clock names"
This reverts commit a71be723fdc4217f3c2a774c151d5be0f452b409.

The MT8186 clock driver can handle automatic reparenting and rate
changes of the GPU PLL and module clock. This used to be handled
manually in the Mali driver. However with the upstream Panfrost driver
it was too platform specific and was deemed to be better handled within
the clock driver.

The Mali driver has been switched over to supporting the upstream
scheme and letting the clock driver handle the reparenting. Drop
the downstream changes.

BUG=b:351768680
UPSTREAM-TASK=b:260376782
TEST=Boot up to login screen

Change-Id: I4d8b377db157ed25ee98682601709f4b50e88ab8
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5948576
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>