CHROMIUM: iwl7000: mld: support MAC cmd for AP
Add the support for sending the MAC cmd in AP mode.

In AP mode, we need to filter in beacons from other APs
for HT protection.
But this is only needed if we are TXing, on the other hand, receiving
every beacon is causing a lot of unnecessary wakeups.
Hold a counter of the associated STAs, so we enable beacons only when
there is at least one STA associated (before that there is no TX)

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I46733343e8ad1d7412c1de69094aeaed0753d73b
iwl7000-tree: 1b2f34ef48c85cc40ead26442090354f6f790ae7