CHROMIUM: iwl7000: fw: read STEP table from correct UEFI var
This variable exists for the "common" (WiFi/BT) GUID, not the
WiFi-only GUID. Fix that by passing the GUID to the function.
A short-cut for the wifi-only version remains so not all code
must be updated.

However, rename the GUID defines to be clearer.

Change-Id: Idae95a70ff69d2ba1b610e8eced826961ce7de98
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 18ec99d6abb8cfc832924448803d56e12a3100f5