CHROMIUM: iwl7000: mld: register as thermal zone
Add the required code to register to the thermal zone subsystem.
Provide the required handler to measure the temperature and to configure
the thresholds in the firmware.
Since we may have several devices being enumerated concurrently, add an
atomic static variable to determine the index of the thermal device.

Also add the bitops header file to the fw/api/phy.h which relies on it.

Change-Id: I34baec7201964551f0c47dce0dbd71a9d7e1b430
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 2d85e56559487cc186b011bc15af6845e50d8363