CHROMIUM: iwl7000: mac80211: Refactor ieee80211_sta_wmm_params()
The function first updates the link configuration and then
calls the driver to set the link parameters. Since the call
to the driver might sleep, split the function such that
the link configuration could be done without calling the
driver. This would be useful in cases that WMM parameters
need to be configured, but the current locking doesn't allow
to call the driver.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Change-Id: I65783c102d44127035838f97fab64ec4df5c40f3
iwl7000-tree: d6bc181b273ad96addce5198f369ce804d78c8e7