CHROMIUM: iwl7000: mld: fix 6 GHz scan construction
If more than 255 colocated APs exist for the set of all
APs found during 2.5/5 GHz scanning, then the 6 GHz scan
construction will loop forever since the loop variable
has type u8, which can never reach the number found when
that's bigger than 255, and is stored in a u32 variable.
Also move it into the loops to have a smaller scope.

Using a u32 there is fine, we limit the number of APs in
the scan list and each has a limit on the number of RNR
entries due to the frame size. With a limit of 1000 scan
results, a frame size upper bound of 4096 (really it's
more like ~2300) and a TBTT entry size of at least 11,
we get an upper bound for the number of ~372k, well in
the bounds of a u32.

Signed-off-by: Daniel Gabay <daniel.gabay@intel.com>
Change-Id: I9af4801d81f692eb87828e835a6240ba5fad2117
iwl7000-tree: 7018880e135bf1ec73f0cca94187c5637183d5a0