CHROMIUM: iwl7000: make IWL_FW_CHECK_FAILED() usable for trans
The non-debug version of the macro already is usable with
a "struct iwl_trans *" argument, but for the debug version
we can't do that because trans->trans doesn't work. Add a
workaround for this using _Generic, which itself needs a
workaround with a nested _Generic to make it work.

Change-Id: I943a895743f5c566a79da89509a8d35420a89702
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 9f879c80d87a3479946d3660c91b5c3f6597e4b6