CHROMIUM: iwl7000: mld: iwl_mld_deactive_link shouldn't fail
If an error ocurred just before or when sending the link command to
deactivate a link, sending the command will fail, and therefore the link
will still be marked as active.
Later when the vif is removed, we will hit the warning about removing an
active link.
On the other hand, iwl_mld_deactive_link is called by
iwl_mld_unassign_vif_chanctx which is void, so mac80211 will not even be
aware of we failing to deactivate the link.
Make iwl_mld_deactive_link void, and mark the link as inactive even if
sending the commans failed.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I20d96183b24b2159a19c15bbae9d60d48ef67a73
iwl7000-tree: 0bc1e3bc468387fab944e14af94c8a6f85387545