CHROMIUM: mm/mglru: discount potential refaults by gen size
When deciding whether to evict file or anon pages, one factor is the
probability that an evicted page of a given type will be refaulted.
However, evicting a page of type A cannot prevent the refault of a page
of type B if that refault would occur after more refaults than there are
pages of type A - e.g. if there are only 100MB of anonymous memory, then
evicting an anonymous page cannot prevent a file refault that will occur
after 1GB of file refaults. As such, the estimated refault probability
needs to be adjusted by the fraction of evictable pages of the opposite
type.

This change addresses an issue where when there is very large imbalance
between the two page types, generations of the smaller type don't live
for long enough to generate refaults and push back against the other
page type.

BUG=b:342531825
UPSTREAM-TASK=b:357764239
TEST=tast run DUT multivm.Lifecycle.host_with_bg_arc_vmmms

Change-Id: I610a690efeceb37aa67de9a6246bc0f4dadcbe67
Signed-off-by: David Stevens <stevensd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6009095
Reviewed-by: Suleiman Souhlal <suleiman@google.com>