CHROMIUM: usb: typec: ucsi: Only check partner PDOs for DRP
ucsi_get_pd_caps checks returned PDOs to determine if the port has a DRP
partner connected. But, on init this function is used to check the port
PDOs. If the port is a DRP, then this function will set the DRP partner
bool regardless of the partner's PDOs.

Update ucsi_get_pd_caps to only check partner PDOs when determining if a
DRP partner is connected.

BUG=b:376619155
UPSTREAM-TASK=b:353605100
TEST=Manually checked Lotso didn't mistake an source-only charger for
a DRP on boot.

Change-Id: If81bbbece3cd7a5f39528e6da3eb4d1a2738d4d8
Signed-off-by: Jameson Thies <jthies@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6013424
Reviewed-by: Benson Leung <bleung@google.com>