CHROMIUM: iwl7000: limit printed string from FW file
There's no guarantee here that the file is always with a
NUL-termination, so reading the string may read beyond the
end of the TLV. If that's the last TLV in the file, it can
perhaps even read beyond the end of the file buffer.

Fix that by limiting the print format to the size of the
buffer we have.

Change-Id: Idec695d53c6c2234aade306f7647b576c7e3d928
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 3ad6b476549a2e1db701867433e248c057ad8deb