CHROMIUM: iwl7000: mld: add an indicator for session_protection request
A command to cancel a session protection is only sent if we have an
indication for an active session. That indication is the session's end time
passed by the firmware with the SESSION_PROTECTION_NOTIF.

A race can occur between the handling of the notification and a demand
for canceling the session. Add an indication for a requested
session protection which wasn't yet answered to determine whether we
should cancel a session.

Signed-off-by: yedidyab <yedidya.ben.shimol@intel.com>
Change-Id: Ie225e995c4226d7361bb352c5f1aa6a65dc9c9fa
iwl7000-tree: dfdb450dbe0dfb37778ec089bf4eff1416595ebd