CHROMIUM: iwl7000: mac80211: log link information in ieee80211_determine_chan_mode
ieee80211_determine_chan_mode is called for each link and if there is a
downgrade, then it is interesting to know on which link it happened.
Pass through the link_id where relevant and use the new link_id_info
macro instead of sdata_info so that the link ID is printed when
relevant.

Change-Id: I64775ec914603d3c7b0c6ea14b507c0370c11622
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: 635ed1395c79a1aae039af9c9789391ebb45ef85