CHROMIUM: iwl7000: mvm: fix add stream vs. restart race
My recent restart related work has made this race more likely
to happen and we've now noticed it, but it seems that it was
always possible. The race is that the add stream work can be
scheduled just before a restart is scheduled and then execute
before the restart, accessing the device while it's doing the
restart and not accessible.

To fix this, check if the device is restarting and abort the
work in that case. Reschedule it after the restart as well.

Change-Id: I1a32936776f8ba5e83dda0a68ffc2722d9d37950
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: fc024a92ced87b498d90ee9cb46c3207ca115078