CHROMIUM: iwl7000: mld: fix add TXQ vs. restart race
My recent restart related work has made this race more likely
to happen and we've now noticed it, but it seems that it was
always possible. The race is that the add TXQ work can be
scheduled just before a restart is scheduled and then execute
before the restart, accessing the device while it's doing the
restart and not accessible.

To fix this, check if the device is restarting and abort the
work in that case. Reschedule it after the restart as well.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I1930f5ae01da99db3ffaf27bfb5e81107bccd32a
iwl7000-tree: f64215241661dfc16c52ebe3533fef10b321fc74