CHROMIUM: net: wwan: t7xx Fix remove process in fastboot state
When the module is damaged, it automatically enters recovery
mode and boots into fastboot state. If the module is switched
to fastboot again via devlink in this state, the remove process
does not account for this scenario, leading to issues in the
exit interaction logic for normal mode.

UPSTREAM-TASK=b:302351783
BUG=b:347768151
TEST=Trigger module damaged and verify the modemfwd recovery

Signed-off-by: jack song <jinjian.song@fibocom.corp-partner.google.com>
Signed-off-by: liuqf <liuqf@fibocom.corp-partner.google.com>
Change-Id: I5b2664497cc132f08c736746f3264877701e5d93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6023564
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Kornel Dulęba <korneld@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>