CHROMIUM: iwl7000: mld: avoid fetching link for pairwise keys
When setting a pairwise key the link ID will be -1. As such, the
iwl_mld_get_key_sta_mask function should not check that the link is
valid in that case.

Move the code into the AP block where it is needed and remove the
fallback to link ID 0.

Change-Id: Idb13388411ef9961d38094cfc715ebeb76e3a29a
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: cb2509035c3b9daa0d56176bc384714f2355a952