CHROMIUM: iwl7000: mvm: avoid NULL pointer dereference
When iterating over the links of a vif, we need to make sure that the
pointer is valid (in other words - that the link exists) before
dereferncing it.
Use for_each_vif_active_link that also does the check.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I7fb7036a0b187c1636b01970207259cb2327952c
iwl7000-tree: 1ad3060728a6d1fdc1ee000cededd8a4dfbc5c56