CHROMIUM: iwl7000: mld: guard against a non-allocated link_sta
If an error happens at the wrong time, then we might be in a situation
where a link station is not allocated even though mac80211 thinks it is.
This could happen if mac80211 does a link switch, but we happen to
receive an FW error during old link. In this case, the new link may not
yet be allocated internally.
Guard against this scenario.

Change-Id: I0d08cd4e1f65f02d136c6c7acf55b5045679027d
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: 421116de5b3006ac16d6c52ef94544e5da60ddb9