CHROMIUM: iwl7000: mld: Fix locking in debugfs write handler
__iwl_dbgfs_##name##_write uses the iwl_mld_<something>_from_mac80211 to
get the driver level representation of the object (vif, link or
link_sta).

But for both link and link_sta, iwl_mld_<something>_from_mac80211 is
dereferencing the relevant pointer with rcu_dereference_check.

We can't have the wiphy lock here, and having an rcu read section is
also wrong.

All we really need it the iwl_mld object, and not the driver level
representation of the link/vif/link_sta. Get it from the mld_vif,
which we can get without rcu_dereference_check from all the objects.

This allows to get rid of the mldtype argument in
_WIPHY_DEBUGFS_WRITE_WRAPPER, and therefore also from
_WIPHY_DEBUGFS_WRITE_FILE_OPS.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ifadf053eadf859b9b361c2700acbe205461f471e
iwl7000-tree: 862237ed16a4df94819e5b361a2065aa92cebdde