CHROMIUM: iwl7000: mac80211: set key link ID to the deflink one
When in non-MLO mode, the key ID was set to -1 even for keys that are
not pairwise. Change the link ID to be the link ID of the deflink in
this case so that drivers do not need to special cases for this.

Change-Id: I4a5c288465e75119edb6a0df90dddf6f30d14a02
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: 754f1ec79ac036d5afa4bca271394f3865aa7cd1