CHROMIUM: iwl7000: mld: cleanup the link also in AP mode
In reconfig, the state of all the interfaces, and of all its links - is
cleaned up.
But this is now only done for STA mode, but the cleanup of the bcast and
mcast stations are needed for AP mode. Do the cleanup also for AP mode.
While at it, allow it also for P2P client/GO

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I64d73882242f57530c8c077de5c062aeb7297138
iwl7000-tree: 8d07fcd0f85cc1fd69d9212bf69a7d334b6f7c3d