CHROMIUM: iwl7000: mld: randomize the firmware resource id
In order to find bugs that are related to the firmware resource id, just
randomize it. If we don't, any test that runs a single interface with a
since link and station will use the index 0 in firmware which can hide
issues that can pop up later in more complex flows.

Since we now use the logger in a file that is included by the kunit
module, import the IWLWIFI namespace.

Change-Id: I1effc00aaed538a9baf4de4b563c59129a65b63a
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 6e857075887723481daa9a0849cd91046daa6d17