CHROMIUM: iwl7000: mld: avoid configuring power during restart
During a restart the links are only re-added properly later on. As such,
it is not possible to configure the power yet which would result in a
warning.

Fundamentally, this is a workaround for the mac80211 behaviour that
first creates the interface and then configures the (initial) links
using a drv_change_vif_links that enables MLO.

Change-Id: Ia06f1d6d419a9c9a64a2738fb5bd564a02a3e4e9
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: 663ca12c43248ff0f24f33fcd1011a3a371aaec8