CHROMIUM: iwl7000: mld: avoid infinite loop
When drv_wake_txq is called, and the wanted txq is not allocated yet,
we add it to txqs_to_add list, and schedule add_txqs_wk which will
itertae over the list and allocate all the txqs in it.

If an error occurred after adding an item to the list, but before the wk
got to run, then the wk will run after the reconfig wk
(either because it was scheduled in this order, or because it is deferred
to after the reconfig).

But mld_txq::list is in the zeroed_on_hw_restart part of the struct,
which means that it will be reset upon the restart cleanup,
and then reinitialized when iwl_mld_init_txq is called
(when the sta is re-added),
so the list gets corrupted such that list_empty is never false.

Then we keep allocating txqs until the FW asserts that there is no more
space for tx queues.

Fix this by preserving mld_txq::list during restart.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I15e38d29c39cef70c83d79534454c117c808ea9a
iwl7000-tree: 955dcfface6c5d7b3aa100862c011a9c8b01ee46