CHROMIUM: iwl7000: mld: remove unnecessary ap_sta assignment
The ap_sta is now already NULL here due to pre-RCU remove
clearing it. As a result, we no longer need to set it to
NULL here, which is also wrong for TDLS, and unnecessary
for AP mode.

Change-Id: I58880d4f2979254365a203057a1a0809dab47447
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 9b7368f8c6b9090adda1a196cc425997a1aac3bc