CHROMIUM: sched: Fix the dl_server reset logic
dl_server field in task_struct should only be set when the task is
selected by dlserver. There were couple of code path where the reset was
not happening and this caused incorrect usage of dlserver. There are
cases where dl_server was not reset and task migrates to another cpu.
One such example caused triggering of the following WARNING:

 WARNING: CPU: 0 PID: 2682 at kernel/sched/sched.h:1377 update_curr_dl_se+0x4b8/0x6f1
 <TASK>
  update_curr+0x12f/0x327 kernel/sched/fair.c:980
  put_prev_entity+0x3c/0x139 kernel/sched/fair.c:5123
  put_prev_task_fair+0x39/0x5f kernel/sched/fair.c:8132
  put_prev_task kernel/sched/sched.h:2269 [inline]
  put_prev_task_balance+0x12d/0x190 kernel/sched/core.c:6006
  __pick_next_task kernel/sched/core.c:6063 [inline]
  pick_next_task kernel/sched/core.c:6124 [inline]
  __schedule+0x970/0x43a0 kernel/sched/core.c:6703
  preempt_schedule_common+0x74/0xc0 kernel/sched/core.c:6920
  preempt_schedule+0xd6/0xdd kernel/sched/core.c:6944
  preempt_schedule_thunk+0x16/0x18 arch/x86/entry/thunk_64.S:34
  __raw_spin_unlock_irq include/linux/spinlock_api_smp.h:160 [inline]
  _raw_spin_unlock_irq+0x3c/0x3e kernel/locking/spinlock.c:202
  spin_unlock_irq include/linux/spinlock.h:401 [inline]
  call_usermodehelper_exec_async+0x7a/0x336 kernel/umh.c:75
  ret_from_fork+0x1f/0x30 arch/x86/entry/entry_64.S:295
 </TASK>

Fix those paths where dl_server field was not reset correctly.

Upstream has taken care of this by a small rework in dlserver:
commit bd9bbc96e835 ("sched: Rework dl_server")

Backporting is not trivial as this commit is based on a set of patches
for sched_ext preparation.

BUG=b:376833598
TEST=syzkaller repro program doesn't crash

Change-Id: I2a5015fe8d26d8f684e4c8cf7fa11c1eebb0f643
Signed-off-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6051060
Reviewed-by: Suleiman Souhlal <suleiman@google.com>