CHROMIUM: iwl7000: mld: only call early key code for group keys
The AP early key handling code is per-link as it is not used for
pairwise keys. Stop calling the function for pairwise keys as they do
not have a link and falling back to the link zero is not valid in an MLO
setup.

Change the code a bit to ensure this invariant holds instead of falling
back to link zero. This relies on mac80211 passing a zero link ID in a
non-MLO setup.

Change-Id: If5d48e535c5acc5ee6bd9a590a845bbcd123ccf2
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: b0ffdde653a4bb754f0a74f05acdddfb48e0732d