CHROMIUM: media: platform: mediatek: isp: Fix imgsys crash issue
Once the user receives the done event, they immediately initialize and
use the media request. However, on the other hand, the kernel has not
released all resources after sending the event. When the user resets
the media request and returns to the kernel to continue with the
incomplete release flow, it leads to accessing released resources,
which causes a crash.

BUG=b:359451218
TEST=emerge-geralt sys-kernel/chromeos-kernel-6_1
UPSTREAM-TASK=b:297128621

Change-Id: Ib694be808ddddba58a30eb53dfe6d0d5a82cff4e
Signed-off-by: Olivia Wen <olivia.wen@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6034177
Reviewed-by: Yunke Cao <yunkec@chromium.org>
Commit-Queue: Harvey Yang <chenghaoyang@chromium.org>
Reviewed-by: Harvey Yang <chenghaoyang@chromium.org>