CHROMIUM: iwl7000: mld: allow not-random FW id for debug
In some cases we need to be able to predict the ID of the different
objects. For example when we want to inject a notification that has an
object id.
Add a debug option to disable the random FW id allocation.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: If12a0ebb52d1b1b8a7d41a43f59127064df173b3
iwl7000-tree: e5b5a4cacfb957d8d626a0f6391906ae180ac00c