CHROMIUM: iwl7000: mld: don't remove keys that weren't added
WEP, and TKIP on AP mode, is done in sw crypto, which means that we are
not sending those keys to the FW.
mac80211 is not supposed to tell us to remove keys that we refused to
add, and it doesn't, but in STA mode, we are iterating all the keys to
make sure we remove the group keys on time (before removing the AP sta)

This iterator was supposed to skip WEP/TKIP keys in this condition:
if (key->hw_key_idx == STA_KEY_IDX_INVALID)
	return;

But those keys were never marked as invalid (i.e. not added to FW)
So we remove them, and then the FW asserts.

Fix this by marking such keys as INVALID.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I8b19a39dbb0c10f2c5c86afeef826828dbfc6be6
iwl7000-tree: f781fc34a289ce4c135bdd132c658dde382d689c